Housekeeping, minor styling fixes. Created contacts now have a random lat and long.
