+++ src/App.css
@@ -15,7 +15,7 @@
 }
 
 .menu {
-    flex: 0 0 200px; /* Fixed width for the menu */
+    flex: 0 0 200px;
     border-right: 1px solid gray;
     padding: 20px;
     display: flex;
@@ -24,13 +24,16 @@
 }
 
 #contacts-list {
-    max-height: 90vh;
+    flex: 1;
     overflow-y: auto;
 }
 
 .main-content {
+    display: flex;
+    flex-direction: column;
     flex: 1;
     padding: 20px;
+    overflow: hidden;
 }
 
 button.delete {
@@ -54,6 +57,7 @@ button.delete {
 .buttons {
     display: flex;
     gap: 4px;
+    padding-top: 10px;
 }
 
 form {
@@ -67,9 +71,10 @@ input {
     padding: 4px;
     max-width: 400px;
     border-radius: 5px;
-    border: 1px solid #a7a7a7; /* Subtle border */
-    box-shadow: inset 0 1px 2px rgba(0, 0, 0, 0.1); /* Subtle shadow */
+    border: 1px solid #a7a7a7; 
+    box-shadow: inset 0 1px 2px rgba(0, 0, 0, 0.1); 
 }
+
 .details-link {
     padding-right: 1.5rem;
 }
