+++ src/components/main/CreateContact.jsx
@@ -3,6 +3,7 @@ import { useNavigate } from "react-router-dom";
 
 import { ApiBase, ContactsContext } from "../../App";
 import MainContent from "./MainContent";
+import ContactForm from "./ContactForm";
 
 const initialData = {
   firstName: "",
@@ -18,28 +19,17 @@ function CreateContact() {
   const { updateContacts } = useContext(ContactsContext);
   const navigate = useNavigate();
 
-  const handleChange = (event) => {
-    const { name, value, type } = event.target;
-
-    if (type === "checkbox") {
-      let tempData = formData[name].filter((v) => v != value);
-      if (event.target.checked) {
-        tempData.push(value);
-      }
-      setFormData({ ...formData, [name]: tempData });
-    } else {
-      setFormData({ ...formData, [event.target.name]: event.target.value });
-    }
-  };
-
   const handleSubmit = (event) => {
     event.preventDefault();
+    var reqData = { ...formData };
+    reqData.latitude = Math.random() * 180 - 90;
+    reqData.longitude = Math.random() * 360 - 180;
     fetch(ApiBase, {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
       },
-      body: JSON.stringify(formData),
+      body: JSON.stringify(reqData),
     })
       .then((res) => {
         if (!res) {
@@ -56,51 +46,12 @@ function CreateContact() {
 
   return (
     <MainContent title="Creating contact:">
-      <form onSubmit={handleSubmit}>
-        <label htmlFor="firstName">First name:</label>
-        <input
-          type="text"
-          name="firstName"
-          value={formData.firstName}
-          onChange={handleChange}
-        />
-        <label htmlFor="lastName">Last name:</label>
-        <input
-          type="text"
-          name="lastName"
-          value={formData.lastName}
-          onChange={handleChange}
-        />
-        <label htmlFor="gender">Gender:</label>
-        <input
-          type="text"
-          name="gender"
-          value={formData.gender}
-          onChange={handleChange}
-        />
-        <label htmlFor="jobTitle">Job Title:</label>
-        <input
-          type="text"
-          name="jobTitle"
-          value={formData.jobTitle}
-          onChange={handleChange}
-        />
-        <label htmlFor="street">Street:</label>
-        <input
-          type="text"
-          name="street"
-          value={formData.street}
-          onChange={handleChange}
-        />
-        <label htmlFor="city">City:</label>
-        <input
-          type="text"
-          name="city"
-          value={formData.city}
-          onChange={handleChange}
-        />
-        <input className="fancy" type="submit" value="Create contact" />
-      </form>
+      <ContactForm
+        handleSubmit={handleSubmit}
+        formData={formData}
+        setFormData={setFormData}
+        submitText={"Create Contact"}
+      />
     </MainContent>
   );
 }
