+++ src/components/main/EditContact.jsx
@@ -3,11 +3,13 @@ import { useNavigate, useParams } from "react-router-dom";
 
 import { ApiBase, ContactsContext } from "../../App";
 import MainContent from "./MainContent";
+import ContactForm from "./ContactForm";
 
 const initialData = {
   firstName: "",
   lastName: "",
   gender: "",
+  jobTitle: "",
   street: "",
   city: "",
 };
@@ -24,26 +26,13 @@ function EditContact() {
       firstName: contact.firstName,
       lastName: contact.lastName,
       gender: contact.gender,
+      jobTitle: contact.jobTitle,
       street: contact.street,
       city: contact.city,
     });
   }, [contact]);
   if (contact == null) return <p>Failed to load contact...</p>;
   const fullName = `${contact.firstName} ${contact.lastName}`;
-
-  const handleChange = (event) => {
-    const { name, value, type } = event.target;
-
-    if (type === "checkbox") {
-      let tempData = formData[name].filter((v) => v != value);
-      if (event.target.checked) {
-        tempData.push(value);
-      }
-      setFormData({ ...formData, [name]: tempData });
-    } else {
-      setFormData({ ...formData, [event.target.name]: event.target.value });
-    }
-  };
 
   const handleSubmit = (event) => {
     event.preventDefault();
@@ -64,50 +53,20 @@ function EditContact() {
         console.log("Updated contact: ", data);
         updateContacts();
         navigate("/contact/" + id);
+      })
+      .catch((error) => {
+        console.error(error);
       });
   };
 
   return (
     <MainContent title={`Editing: "${fullName}"`}>
-      <form onSubmit={handleSubmit}>
-        <label htmlFor="firstName">First name:</label>
-        <input
-          type="text"
-          name="firstName"
-          value={formData.firstName}
-          onChange={handleChange}
-        />
-        <label htmlFor="lastName">Last name:</label>
-        <input
-          type="text"
-          name="lastName"
-          value={formData.lastName}
-          onChange={handleChange}
-        />
-        <label htmlFor="gender">Gender:</label>
-        <input
-          type="text"
-          name="gender"
-          value={formData.gender}
-          onChange={handleChange}
-        />
-
-        <label htmlFor="street">Street:</label>
-        <input
-          type="text"
-          name="street"
-          value={formData.street}
-          onChange={handleChange}
-        />
-        <label htmlFor="city">City:</label>
-        <input
-          type="text"
-          name="city"
-          value={formData.city}
-          onChange={handleChange}
-        />
-        <input className="fancy" type="submit" value="Save changes" />
-      </form>
+      <ContactForm
+        handleSubmit={handleSubmit}
+        formData={formData}
+        setFormData={setFormData}
+        submitText={"Edit Contact"}
+      />
     </MainContent>
   );
 }
